Fixed possible int overflow because "day" might be bigger than maximum possible int.

+++ src/main/java/com/smart4aviation/queries/TotalCapacityByPlanesActiveOnDay.java
@@ -3,9 +3,9 @@ package com.smart4aviation.queries;
 public class TotalCapacityByPlanesActiveOnDay implements Query {
     private int firstRoute;
     private int lastRoute;
-    private int day;
+    private long day;
 
-    public TotalCapacityByPlanesActiveOnDay(int firstRoute, int lastRoute, int day) {
+    public TotalCapacityByPlanesActiveOnDay(int firstRoute, int lastRoute, long day) {
         this.firstRoute = firstRoute;
         this.lastRoute = lastRoute;
         this.day = day;
